% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr_output}
\alias{slr_output}
\title{Simple Linear Regression}
\usage{
slr_output(x, y)
}
\arguments{
\item{x}{The predictor variable.}

\item{y}{The outcome variable.}
}
\value{
The intercept and slope estimates, standard error of slope, f statistic, t statistic of slope, and corresponding p-values.
}
\description{
Used to carry out simple linear regression.
}
\examples{
data(iris)
slr_output(iris$Sepal.Width, iris$Sepal.Length)
data(mtcars)
slr_output(mtcars$mpg, mtcars$wt)
}
