% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr.R
\name{slr_output}
\alias{slr_output}
\title{Simple Linear Regression}
\usage{
slr_output(x, y)
}
\arguments{
\item{x}{The predictor variable.}

\item{y}{The outcome variable.}
}
\value{
The intercept and slope estimates, standard error, f and t statistics and p-values.
}
\description{
Used to carry out simple linear regression.
}
\examples{
data(mtcars)
slr_output(mtcars$disp, mtcars$wt)
}
